pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ruslan Shamsutdinov <nexus191@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
Flask = "^2.3.2"
Jinja2 = "^3.1.2"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
